requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-core
langchain-community
langchain-openai
langgraph

chromadb
arxiv

# optional: local LLMs later
ollama


pypdf
unstructured

chromadb
langsmith
streamlit